fix(bake): locate baseami
